+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+unreleased
+-----
+- CO-2544 prevent zip bomb upload during image push
+
 5.2.12
 -----
 - add provide for jtl/connector-mapping-tables and jtl/connector-dbc